+++ plone/app/blocks/resource.py
@@ -1,5 +1,10 @@
 # -*- coding: utf-8 -*-
+from ConfigParser import SafeConfigParser
+import logging
+import urlparse
+
 from Acquisition import aq_parent
+import Globals
 from OFS.interfaces import ITraversable
 from Products.CMFCore.utils import getToolByName
 from plone.app.blocks.interfaces import DEFAULT_AJAX_LAYOUT_REGISTRY_KEY
@@ -15,8 +20,9 @@ from plone.app.blocks.utils import resolveResource
 from plone.memoize import view
 from plone.memoize import volatile
 from plone.registry.interfaces import IRecordModifiedEvent
-from plone.resource.manifest import getAllResources
+from plone.resource.manifest import MANIFEST_FILENAME
 from plone.resource.traversal import ResourceTraverser
+from plone.resource.utils import iterDirectoriesOfType
 from plone.subrequest import ISubRequest
 from zExceptions import NotFound
 from zope.annotation import IAnnotations
@@ -28,8 +34,9 @@ from zope.schema.interfaces import IVocabularyFactory
 from zope.schema.vocabulary import SimpleTerm
 from zope.schema.vocabulary import SimpleVocabulary
 from zope.site.hooks import getSite
-import Globals
-import urlparse
+
+
+logger = logging.getLogger('plone.app.blocks')
 
 
 class SiteLayoutTraverser(ResourceTraverser):
@@ -48,6 +55,81 @@ class AnnotationsDict(dict):
     implements(IAnnotations)
 
 
+class multidict(dict):
+    """
+    Taken from: http://stackoverflow.com/questions/9876059/parsing-configure-file-with-same-section-name-in-python  # noqa
+    """
+    _unique = 0
+
+    def __setitem__(self, key, val):
+        if isinstance(val, dict):
+            self._unique += 1
+            key += str(self._unique)
+        dict.__setitem__(self, key, val)
+
+
+def getLayoutsFromManifest(fp, format, directory_name):
+    parser = SafeConfigParser(None, multidict)
+    parser.readfp(fp)
+
+    layouts = {}
+    for section in parser.sections():
+        if not section.startswith(format.resourceType) or ':variants' in section:
+            continue
+        # id is a combination of directory name + filename
+        if parser.has_option(section, 'file'):
+            filename = parser.get(section, 'file')
+        else:
+            filename = ''  # this should not happen...
+        _id = directory_name + '/' + filename
+        if _id in layouts:
+            # because TTW resources are created first, we consider layouts
+            # with same id in a TTW to be taken before other resources
+            continue
+        data = {
+            'directory': directory_name
+        }
+        for key in format.keys:
+            if parser.has_option(section, key):
+                data[key] = parser.get(section, key)
+            else:
+                data[key] = format.defaults.get(key, None)
+        layouts[_id] = data
+
+    return layouts
+
+
+def getLayoutsFromResources(format):
+    layouts = {}
+
+    for directory in iterDirectoriesOfType(format.resourceType):
+
+        name = directory.__name__
+        if directory.isFile(MANIFEST_FILENAME):
+            manifest = directory.openFile(MANIFEST_FILENAME)
+            try:
+                layouts.update(getLayoutsFromManifest(manifest, format, name))
+            except:
+                logger.exception("Unable to read manifest for theme directory %s", name)
+            finally:
+                manifest.close()
+        else:
+            # can provide default file for it with no manifest
+            filename = format.defaults.get('file', '')
+            if filename and directory.isFile(filename):
+                _id = name + '/' + filename
+                if _id not in layouts:
+                    # not overridden
+                    layouts[_id] = {
+                        'title': name.capitalize().replace('-', ' ').replace('.', ' '),
+                        'description': '',
+                        'directory': name,
+                        'file': format.defaults.get('file', '')
+                    }
+
+    return layouts
+
+
 class _AvailableLayoutsVocabulary(object):
     """Vocabulary to return request cached available layouts of a given type
     """
@@ -61,28 +143,20 @@ class _AvailableLayoutsVocabulary(object):
     def __call__(self, context, format, defaultFilename):
         items = {}  # dictionary is used here to avoid duplicate tokens
 
-        resources = getAllResources(format)
-        for name, manifest in resources.items():
-            title = name.capitalize().replace('-', ' ').replace('.', ' ')
-            filename = defaultFilename
+        resources = getLayoutsFromResources(format)
+        used = []
+        for _id, config in resources.items():
+            title = config.get('title', _id)
+            filename = config.get('file', defaultFilename)
 
-            if manifest is not None:
-                title = manifest['title'] or title
-                filename = manifest['file'] or filename
-                variants = manifest.get('variants') or []
-            else:
-                variants = []
-
-            path = "/++%s++%s/%s" % (format.resourceType, name, filename)
-            items[name] = SimpleTerm(path, name, title)
-
-            for key, value in dict(variants).items():
-                key_ = title_ = key.capitalize().replace('_', ' ')
-                name_ = '{0:s}-{1:s}'.format(name, key)
-                if manifest is not None and manifest['title']:
-                    title_ = u'{0:s} ({1:s})'.format(title, key_)
-                path = "/++%s++%s/%s" % (format.resourceType, name, value)
-                items[name_] = SimpleTerm(path, name_, title_)
+            path = "/++%s++%s/%s" % (format.resourceType, config['directory'], filename)
+            if path in used:
+                # term values also need to be unique
+                # this should not happen but it's possible for users to screw up
+                # their layout definitions and it's better to not error here
+                continue
+            used.append(path)
+            items[_id] = SimpleTerm(path, _id, title)
 
         items = sorted(items.values(), key=lambda term: term.title)
         return SimpleVocabulary(items)